+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.19.1]
+
+ * Make EventTimeConverter serialize without the need for an attribute [#51]
+
+## [0.19.0]
+
+ * Make EventTimeConverter public [#50]
+
 ## [0.18.0]
 
  * Support returning Google event IDs [#48]
@@ -123,6 +131,8 @@
 [0.16.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.16.0
 [0.17.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.17.0
 [0.18.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.18.0
+[0.19.0]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.19.0
+[0.19.1]: https://github.com/cronofy/cronofy-csharp/releases/tag/rel-0.19.1
 
 [#3]: https://github.com/cronofy/cronofy-csharp/pull/3
 [#10]: https://github.com/cronofy/cronofy-csharp/pull/10
@@ -152,3 +162,5 @@
 [#46]: https://github.com/cronofy/cronofy-csharp/pull/46
 [#47]: https://github.com/cronofy/cronofy-csharp/pull/47
 [#48]: https://github.com/cronofy/cronofy-csharp/pull/48
+[#50]: https://github.com/cronofy/cronofy-csharp/pull/50
+[#51]: https://github.com/cronofy/cronofy-csharp/pull/51